添加了动态的图片的显示和头像的显示

+++ app/src/main/java/com/coder/zt/sobblog/utils/ScreenUtils.kt
@@ -10,7 +10,7 @@ class ScreenUtils {
         /**
          * 根据手机的分辨率从 dp 的单位 转成为 px(像素)
          */
-        fun dp2px(dpValue: Float): Int {
+        fun dp2px(dpValue: Int): Int {
             return (0.5f + dpValue * Resources.getSystem().displayMetrics.density).toInt()
         }
     }
